+++ sync-for-payroll/CodatSyncPayroll/Companies.cs
@@ -81,12 +81,12 @@ namespace Codat.Sync.Payroll
     /// </summary>
     public class Companies: ICompanies
     {
-        public SDKConfig Config { get; private set; }
+        public SDKConfig SDKConfiguration { get; private set; }
         private const string _language = "csharp";
-        private const string _sdkVersion = "3.0.0";
-        private const string _sdkGenVersion = "2.195.2";
+        private const string _sdkVersion = "3.0.1";
+        private const string _sdkGenVersion = "2.209.0";
         private const string _openapiDocVersion = "3.0.0";
-        private const string _userAgent = "speakeasy-sdk/csharp 3.0.0 2.195.2 3.0.0 Codat.Sync.Payroll";
+        private const string _userAgent = "speakeasy-sdk/csharp 3.0.1 2.209.0 3.0.0 Codat.Sync.Payroll";
         private string _serverUrl = "";
         private ISpeakeasyHttpClient _defaultClient;
         private ISpeakeasyHttpClient _securityClient;
@@ -96,13 +96,13 @@ namespace Codat.Sync.Payroll
             _defaultClient = defaultClient;
             _securityClient = securityClient;
             _serverUrl = serverUrl;
-            Config = config;
+            SDKConfiguration = config;
         }
         
 
         public async Task<CreateCompanyResponse> CreateAsync(CompanyRequestBody? request = null)
         {
-            string baseUrl = this.Config.GetTemplatedServerDetails();
+            string baseUrl = this.SDKConfiguration.GetTemplatedServerDetails();
             var urlString = baseUrl + "/companies";
             
             var httpRequest = new HttpRequestMessage(HttpMethod.Post, urlString);
@@ -151,7 +151,7 @@ namespace Codat.Sync.Payroll
 
         public async Task<DeleteCompanyResponse> DeleteAsync(DeleteCompanyRequest? request = null)
         {
-            string baseUrl = this.Config.GetTemplatedServerDetails();
+            string baseUrl = this.SDKConfiguration.GetTemplatedServerDetails();
             var urlString = URLBuilder.Build(baseUrl, "/companies/{companyId}", request);
             
             var httpRequest = new HttpRequestMessage(HttpMethod.Delete, urlString);
@@ -191,7 +191,7 @@ namespace Codat.Sync.Payroll
 
         public async Task<GetCompanyResponse> GetAsync(GetCompanyRequest? request = null)
         {
-            string baseUrl = this.Config.GetTemplatedServerDetails();
+            string baseUrl = this.SDKConfiguration.GetTemplatedServerDetails();
             var urlString = URLBuilder.Build(baseUrl, "/companies/{companyId}", request);
             
             var httpRequest = new HttpRequestMessage(HttpMethod.Get, urlString);
@@ -235,7 +235,7 @@ namespace Codat.Sync.Payroll
 
         public async Task<ListCompaniesResponse> ListAsync(ListCompaniesRequest? request = null)
         {
-            string baseUrl = this.Config.GetTemplatedServerDetails();
+            string baseUrl = this.SDKConfiguration.GetTemplatedServerDetails();
             var urlString = URLBuilder.Build(baseUrl, "/companies", request);
             
             var httpRequest = new HttpRequestMessage(HttpMethod.Get, urlString);
@@ -279,7 +279,7 @@ namespace Codat.Sync.Payroll
 
         public async Task<UpdateCompanyResponse> UpdateAsync(UpdateCompanyRequest? request = null)
         {
-            string baseUrl = this.Config.GetTemplatedServerDetails();
+            string baseUrl = this.SDKConfiguration.GetTemplatedServerDetails();
             var urlString = URLBuilder.Build(baseUrl, "/companies/{companyId}", request);
             
             var httpRequest = new HttpRequestMessage(HttpMethod.Put, urlString);