+++ sync-for-payroll/CodatSyncPayroll/Journals.cs
@@ -97,12 +97,12 @@ namespace Codat.Sync.Payroll
     /// </summary>
     public class Journals: IJournals
     {
-        public SDKConfig Config { get; private set; }
+        public SDKConfig SDKConfiguration { get; private set; }
         private const string _language = "csharp";
-        private const string _sdkVersion = "3.0.0";
-        private const string _sdkGenVersion = "2.195.2";
+        private const string _sdkVersion = "3.0.1";
+        private const string _sdkGenVersion = "2.209.0";
         private const string _openapiDocVersion = "3.0.0";
-        private const string _userAgent = "speakeasy-sdk/csharp 3.0.0 2.195.2 3.0.0 Codat.Sync.Payroll";
+        private const string _userAgent = "speakeasy-sdk/csharp 3.0.1 2.209.0 3.0.0 Codat.Sync.Payroll";
         private string _serverUrl = "";
         private ISpeakeasyHttpClient _defaultClient;
         private ISpeakeasyHttpClient _securityClient;
@@ -112,13 +112,13 @@ namespace Codat.Sync.Payroll
             _defaultClient = defaultClient;
             _securityClient = securityClient;
             _serverUrl = serverUrl;
-            Config = config;
+            SDKConfiguration = config;
         }
         
 
         public async Task<Models.Operations.CreateJournalResponse> CreateAsync(CreateJournalRequest? request = null)
         {
-            string baseUrl = this.Config.GetTemplatedServerDetails();
+            string baseUrl = this.SDKConfiguration.GetTemplatedServerDetails();
             var urlString = URLBuilder.Build(baseUrl, "/companies/{companyId}/connections/{connectionId}/push/journals", request);
             
             var httpRequest = new HttpRequestMessage(HttpMethod.Post, urlString);
@@ -167,7 +167,7 @@ namespace Codat.Sync.Payroll
 
         public async Task<GetJournalResponse> GetAsync(GetJournalRequest? request = null)
         {
-            string baseUrl = this.Config.GetTemplatedServerDetails();
+            string baseUrl = this.SDKConfiguration.GetTemplatedServerDetails();
             var urlString = URLBuilder.Build(baseUrl, "/companies/{companyId}/data/journals/{journalId}", request);
             
             var httpRequest = new HttpRequestMessage(HttpMethod.Get, urlString);
@@ -211,7 +211,7 @@ namespace Codat.Sync.Payroll
 
         public async Task<GetCreateJournalModelResponse> GetCreateModelAsync(GetCreateJournalModelRequest? request = null)
         {
-            string baseUrl = this.Config.GetTemplatedServerDetails();
+            string baseUrl = this.SDKConfiguration.GetTemplatedServerDetails();
             var urlString = URLBuilder.Build(baseUrl, "/companies/{companyId}/connections/{connectionId}/options/journals", request);
             
             var httpRequest = new HttpRequestMessage(HttpMethod.Get, urlString);
@@ -255,7 +255,7 @@ namespace Codat.Sync.Payroll
 
         public async Task<ListJournalsResponse> ListAsync(ListJournalsRequest? request = null)
         {
-            string baseUrl = this.Config.GetTemplatedServerDetails();
+            string baseUrl = this.SDKConfiguration.GetTemplatedServerDetails();
             var urlString = URLBuilder.Build(baseUrl, "/companies/{companyId}/data/journals", request);
             
             var httpRequest = new HttpRequestMessage(HttpMethod.Get, urlString);